separate iterator into a new file + include diagram
